Create program 5.py
Write a C program for generalization of the Caesar cipher, known as the affine Caesar cipher, has the
following form: For each plaintext letter p, substitute the ciphertext letter C: C = E([a, b], p) = (ap + b) 
mod 26 A basic requirement of any encryption algorithm is that it be one-to-one. That is, if p q, then 
E(k, p) E(k, q). Otherwise, decryption is impossible, because more than one plaintext character maps into 
the same ciphertext character. The affine Caesar cipher is not one-to-one for all values of a. For example, 
for a = 2 and b = 3, then E([a, b], 0) = E([a, b], 13) = 3.
a. Are there any limitations on the value of b? 
b. Determine which values of a are not allowed.